Relocate crypto interface header as separate library
