+++ src/bsl/cose_sc/CoseContext.c
@@ -34,7 +34,7 @@
 #include <m-bptree.h>
 
 #include <bsl/BPSecLib_Private.h>
-#include <bsl/CryptoInterface.h>
+#include <bsl/crypto/CryptoInterface.h>
 #include <bsl/dynamic/CBOR.h>
 
 #include "CoseContext.h"
