+++ src/bsl/crypto/CryptoInterface.h
@@ -29,8 +29,11 @@
 
 #include <stdint.h>
 
-#include "BPSecLib_Private.h"
-#include "BPSecLib_Public.h"
+#include "bsl/BSLConfig.h"
+#include "bsl/front/Data.h"
+#include "bsl/front/SeqReader.h"
+#include "bsl/front/SeqWriter.h"
+#include "bsl/BPSecLib_Private.h" // TODO replace with Variant.h
 
 #ifdef __cplusplus
 extern "C" {
