+++ src/bsl/default_sc/BCB_AES_GCM.c
@@ -31,7 +31,7 @@
 #include <qcbor/qcbor_spiffy_decode.h>
 
 #include <bsl/BPSecLib_Private.h>
-#include <bsl/CryptoInterface.h>
+#include <bsl/crypto/CryptoInterface.h>
 #include <bsl/dynamic/CBOR.h>
 
 #include "DefaultSecContext.h"
