+++ src/bsl/mock_bpa/mock_bpa.c
@@ -35,7 +35,7 @@
 
 #include <bsl/BPSecLib_Private.h>
 #include <bsl/BPSecLib_Public.h>
-#include <bsl/CryptoInterface.h>
+#include <bsl/crypto/CryptoInterface.h>
 #include <bsl/sample_pp/PolicyParser.h>
 
 #include "agent.h"
