+++ test/DefaultScUtils.c
@@ -25,7 +25,7 @@
 #include <m-string.h>
 
 #include <bsl/BPSecLib_Private.h>
-#include <bsl/CryptoInterface.h>
+#include <bsl/crypto/CryptoInterface.h>
 #include <bsl/mock_bpa/MockBPA.h>
 
 #include <bsl/dynamic/IdValPair.h>
