+++ test/DefaultScUtils.h
@@ -24,7 +24,7 @@
 
 #include <m-string.h>
 
-#include <bsl/CryptoInterface.h>
+#include <bsl/crypto/CryptoInterface.h>
 #include <bsl/dynamic/PublicInterfaceImpl.h>
 #include <bsl/dynamic/SecOperation.h>
 #include <bsl/dynamic/IdValPair.h>
