+++ test/fuzz_mock_bpa_keyregistry_cosekey.cpp
@@ -26,7 +26,7 @@
 #include "TestUtils.h"
 #include <bsl/mock_bpa/MockBPA.h>
 #include <bsl/mock_bpa/key_registry.h>
-#include <bsl/CryptoInterface.h>
+#include <bsl/crypto/CryptoInterface.h>
 #include <cinttypes>
 
 #define EXPECT_EQ(expect, got)          \
