+++ test/test_CoseContext.c
@@ -35,7 +35,7 @@
 
 #include <bsl/BPSecLib_Private.h>
 #include <bsl/mock_bpa/MockBPA.h>
-#include <bsl/CryptoInterface.h>
+#include <bsl/crypto/CryptoInterface.h>
 
 #include <bsl/dynamic/PublicInterfaceImpl.h>
 #include <bsl/cose_sc/CoseContext.h>
