+++ test/test_DefaultSecurityContext.c
@@ -36,7 +36,7 @@
 
 #include <bsl/BPSecLib_Private.h>
 #include <bsl/mock_bpa/MockBPA.h>
-#include <bsl/CryptoInterface.h>
+#include <bsl/crypto/CryptoInterface.h>
 
 #include <bsl/dynamic/PublicInterfaceImpl.h>
 #include <bsl/default_sc/DefaultSecContext.h>
